Add file parsing
Adds a data directory to store input/data files for use in the code. Also adds two code libraries, a state and an io. State is the library for all data storage in the code and IO is for input and output operations (currently just the custom file format reader). The custom file format is named the .phys file and it consists of object data needed like mass, initial velocity, position, and more.